+++ .gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 .env
+.ssh
 uploads
 
 # Editor directories and files
