+++ src/styles/index.css
@@ -2,3 +2,7 @@ body {
   margin: 0;
   padding: 0;
 }
+
+h1 {
+  color: var(--primary-color);
+}
